Complete DIMPLE.py module migration
Phase 1 + 2 of the module-migration refactor. The ten functions that the
architecture skeleton had stubbed as re-export shims now own real
implementations in their focused modules:

  - primers.py          find_geneprimer, find_fragment_primer, check_nonspecific
  - fragment_layout.py  recalculate_num_fragments, switch_fragmentsize, check_overhangs
  - oligo_assembly.py   combine_fragments
  - outputs.py          print_all
  - qc.py               post_qc, check_final_assembly

Each module imports the DIMPLE class from the core.py leaf, so the import
graph is a clean DAG: core <- {functional modules} <- DIMPLE.py.

DIMPLE.py is reduced from ~2.3k to ~1.2k lines: it now holds only the
generate_DMS_fragments orchestrator and align_genevariation, and
re-exports the migrated functions at end-of-file so existing
`from DIMPLE.DIMPLE import ...` callers (run_dimple, run_dimple_gui,
tests) are unaffected. The _legacy_* aliases are removed and now-unused
imports (numpy, pydna, MeltingTemp, findORF) pruned.

Also fixed in passing:
  - combine_fragments: the old shim declared a wrong 1-arg signature;
    the real (tandem, num_frag_per_oligo, split) signature is restored.
  - test_qc.py: mock patch target moved to DIMPLE.qc.check_final_assembly
    since post_qc now resolves that call within the qc module.

Full suite green (63 passed).
