+++ DIMPLE/DIMPLE.py
@@ -23,21 +23,15 @@ import re
 import warnings
 from difflib import SequenceMatcher
 from math import ceil
-from DIMPLE.utilities import findORF
 from DIMPLE.core import DIMPLE, addgene
 
-import numpy as np
 from Bio import SeqIO, Align
 from Bio.Seq import Seq
 from Bio.SeqRecord import SeqRecord
-from Bio.SeqUtils import MeltingTemp as mt
 from Bio.SeqUtils import seq3, seq1
 
 import logging
 
-# For testing
-from pydna.dseqrecord import Dseqrecord
-from pydna.amplify import pcr
 from Bio.Restriction import BsmBI, BsaI
 
 logger = logging.getLogger(__name__)
@@ -189,433 +183,6 @@ def align_genevariation(OLS):
         print(
             "No redundant sequences found. Matching sequences may be too short or not aligned to reduce number of oligos synthesized"
         )
-
-def find_geneprimer(genefrag, start, end):
-    # 3' end of primer is variable to adjust melting temperature
-    # 5' end of primer is fixed, with restriction site added
-    # Also add space for maximum deletion on 5' end
-    primer = (
-        genefrag[start:end].complement() + DIMPLE.cutsite[::-1] + "ATA"
-    )  # added ATA for cleavage close to end of DNA fragment
-    # Check melting temperature
-    # find complementary sequences
-    comp = 0  # compensate for bases that align with bsmbi
-    while primer.complement()[end - start + comp] == genefrag[end + comp]:
-        comp += 1
-    # comp += 1 # This is important for single basepair overhang
-    tm2 = mt.Tm_NN(primer[0: end - start + comp], nn_table=mt.DNA_NN2)
-    tm4 = mt.Tm_NN(primer[0: end - start + comp], nn_table=mt.DNA_NN4)
-    count = 0
-    while (
-        tm2 < DIMPLE.gene_primerTm[0]
-        or tm2 > DIMPLE.gene_primerTm[1]
-        or tm4 < DIMPLE.gene_primerTm[0]
-        or tm4 > DIMPLE.gene_primerTm[1]
-    ):
-        if tm2 < DIMPLE.gene_primerTm[0] or tm4 < DIMPLE.gene_primerTm[0]:
-            start += -1
-            primer = (
-                genefrag[start:end].complement() + DIMPLE.cutsite[::-1] + "ATA"
-            )  # cut site addition
-            tm2 = mt.Tm_NN(primer[0: end - start + comp], nn_table=mt.DNA_NN2)
-            tm4 = mt.Tm_NN(primer[0: end - start + comp], nn_table=mt.DNA_NN4)
-        if (
-            count > 12 or start == 0
-        ):  # stop if caught in inf loop or if linker is at max (31 + 7 = 38 bases)
-            break
-        if tm2 > DIMPLE.gene_primerTm[1] and tm4 > DIMPLE.gene_primerTm[1]:
-            start += 1
-            primer = genefrag[start:end].complement() + DIMPLE.cutsite[::-1] + "ATA"
-            # tm = mt.Tm_NN(primer[0:e-s+comp],c_seq=genefrag[s:e+comp],nn_table=mt.DNA_NN2)
-            tm2 = mt.Tm_NN(primer[0: end - start + comp], nn_table=mt.DNA_NN2)
-            tm4 = mt.Tm_NN(primer[0: end - start + comp], nn_table=mt.DNA_NN4)
-        count += 1
-    # optional - force first nucleotide to a C or G
-    # while primer[0]=="T" or primer[0]=="A" or primer[0]=="t" or primer[0]=="a":
-    #     s += -1
-    #     primer = genefrag[s:e].complement()+"CTCTGCA"
-    #     tm = mt.Tm_NN(primer[0:e-s+comp],nn_table=mt.DNA_NN2)
-    # return final primer with tm
-    print(
-        "Generated primers: ",
-        primer.complement().reverse_complement(),
-        round(tm2, 1),
-        start,
-    )
-    return primer.complement().reverse_complement(), round(tm2, 1), start
-
-
-def find_fragment_primer(fragment, stop):
-    # This function finds optimal primer for OLS subpool by changing 3' end of primer
-    start = 0  # starts at maximum length (5' is fixed)
-    if stop > 25:  # limit primer to 25 bases to begin with
-        end = 25
-    else:
-        end = stop
-    count = 0
-    primer = fragment[start:end]
-    tm2 = mt.Tm_NN(
-        primer, nn_table=mt.DNA_NN2
-    )  # Two methods of finding melting temperature seems more consistent
-    tm4 = mt.Tm_NN(primer, nn_table=mt.DNA_NN4)
-    while (
-        tm2 < DIMPLE.primerTm[0]
-        or tm2 > DIMPLE.primerTm[1]
-        or tm4 < DIMPLE.primerTm[0]
-        or tm4 > DIMPLE.primerTm[1]
-        or len(primer) < 16
-    ):
-        count += 1
-        if (
-            count > 12 or end > stop
-        ):  # stop if caught in inf loop or if primer is larger than the barcode
-            end = stop
-            primer = fragment[start:end]
-            break
-
-        if tm2 < DIMPLE.primerTm[0] or tm4 < DIMPLE.primerTm[0]:
-            if start == 0:
-                break
-            end += 1
-            primer = fragment[start:end]
-            tm2 = mt.Tm_NN(primer, nn_table=mt.DNA_NN2)
-            tm4 = mt.Tm_NN(primer, nn_table=mt.DNA_NN4)
-
-        if tm2 > DIMPLE.primerTm[1] or tm4 > DIMPLE.primerTm[1]:
-            end += -1
-            primer = fragment[start:end]
-            tm2 = mt.Tm_NN(primer, nn_table=mt.DNA_NN2)
-            tm4 = mt.Tm_NN(primer, nn_table=mt.DNA_NN4)
-    return primer, round(tm2, 1)
-
-
-def check_nonspecific(primer, fragment, point):
-    non = []
-    # fragment is the entire gene sequence plus the buffer sequence on each side
-    # point is the position of the primer in the fragment
-    # Forward
-    for i in range(len(fragment) - len(primer)):  # Scan each position
-        # first check if the primer binds at each position in the fragment
-        match = [
-            primer[j].lower() == fragment[i + j].lower() for j in range(len(primer))
-        ]
-        first = 10
-        for k in range(len(match) - 3):
-            if (match[k] and match[k + 1] and match[k + 3]) or (
-                match[k] and match[k + 1] and match[k + 2]
-            ):
-                first = k
-                break
-        # if the primer binds to 80% of the first ... bases
-        # and more than 6 bases
-        # and the 3' matches
-        # then check melting temperature
-        if (
-            sum(match[first:]) > len(primer[first:]) * 0.8
-            and sum(match[first:]) > 6
-            and match[-1]
-            and point != i
-        ):  # string compare - sum of matched nt is greater than 80%
-            try:
-                # check the melting temperature of the primer
-                melt = mt.Tm_NN(
-                    primer[first:],
-                    c_seq=fragment[i + first : i + len(primer)].complement(),
-                    nn_table=mt.DNA_NN2,
-                    de_table=mt.DNA_DE1,
-                    imm_table=mt.DNA_IMM1,
-                )
-                if melt > 25:
-                    print("Found non-specific match at " + str(i + 1) + "bp:")
-                    print("match: " + fragment[i: i + len(primer)])
-                    print("primer:" + primer + " Tm:" + str(round(melt, 1)))
-                    logger.warning("Found non-specific match at " + str(i + 1) + "bp:")
-                    logger.warning("match: " + fragment[i: i + len(primer)])
-                    logger.warning("primer:" + primer + " Tm:" + str(round(melt, 1)))
-                if melt > 35:
-                    non.append(True)
-            except ValueError as valerr:
-                print(
-                    str(valerr)
-                    + ". Please check position manually:"
-                    + str(i + 1)
-                    + " forward"
-                )
-                print("Primer:" + primer)
-                print("Match: " + fragment[i : i + len(primer)])
-                non.append(False)
-    # Reverse
-    fragment = fragment.reverse_complement()
-    for i in range(len(fragment) - len(primer)):
-        match = [
-            primer[j].lower() == fragment[i + j].lower() for j in range(len(primer))
-        ]
-        first = 10
-        for k in range(0, len(match) - 3, 1):
-            if match[k] and match[k + 1] and match[k + 3]:
-                first = k
-                break
-        if (
-            sum(match[first:]) > len(primer[first:]) * 0.8
-            and sum(match[first:]) > 6
-            and match[-1]
-            and point != -i
-        ):  # string compare - sum of matched nt is greater than 80%
-            try:
-                melt = mt.Tm_NN(
-                    primer[first:],
-                    c_seq=fragment[i + first : i + len(primer)].complement(),
-                    nn_table=mt.DNA_NN2,
-                    de_table=mt.DNA_DE1,
-                    imm_table=mt.DNA_IMM1,
-                )
-                if melt > 20:
-                    print("Found non-specific match at " + str(i + 1) + "bp:")
-                    print(" match:" + fragment[i : i + len(primer)])
-                    print("primer:" + primer + " Tm:" + str(melt))
-                if melt > 35:
-                    non.append(True)
-            except ValueError as valerr:
-                print(
-                    str(valerr)
-                    + ". Please check position manually:"
-                    + str(i + 1)
-                    + " reverse"
-                )
-                print("Primer:" + primer)
-                print("Match: " + fragment[i : i + len(primer)])
-                non.append(False)
-    return sum(non)
-
-
-def recalculate_num_fragments(gene):
-    num = int(
-        round(((gene.end - gene.start) / float(gene.maxfrag)) + 0.499999999)
-    )  # total bins needed (rounded up)
-    insertionsites = range(gene.start + 3, gene.end, 3)
-    gene.fragsize = [len(insertionsites[i::num]) * 3 for i in list(range(num))]
-    total = DIMPLE.primerBuffer
-    breaksites = [DIMPLE.primerBuffer]
-    for x in gene.fragsize:
-        total += x
-        breaksites.extend([total])
-    # if DIMPLE.dms:
-    #     tmpbreaklist = []
-    #     for idx, x in enumerate(breaksites[:-1]):
-    #         if idx:
-    #             tmpbreaklist.append([x, x + gene.fragsize[idx]])
-    #         else:
-    #             tmpbreaklist.append([x + 3, x + gene.fragsize[idx] + 3])
-    #     gene.breaklist = tmpbreaklist
-    # else:
-    gene.breaklist = [
-        [x, x + gene.fragsize[idx]] for idx, x in enumerate(breaksites[:-1])
-    ]  # insertion site to insertion site
-    #gene.problemsites = set()
-    gene.breaksites = breaksites
-    gene.unique_Frag = [True] * len(gene.fragsize)
-    return gene
-
-
-def switch_fragmentsize(gene, detectedsite, OLS):
-    """TODO:
-    Docstring
-    """
-
-    if not isinstance(gene, DIMPLE):
-        raise TypeError("Not an instance of the DIMPLE class")
-    skip = False
-    count = 0
-    count2 = 0
-    print("Non specific Fragment:" + str(detectedsite))
-    if (
-        len(gene.fragsize) * gene.maxfrag < len(gene.seq) - gene.primerBuffer * 2
-    ):  # if the maxfrag has changed and it is impossible to split the gene into x number of fragments it should recalculate the number of fragments
-        gene = recalculate_num_fragments(gene)
-    else:
-        gene.problemsites.add(gene.breaksites[detectedsite])
-    if all(item == gene.maxfrag for item in gene.fragsize) or any(
-        item > gene.maxfrag for item in gene.fragsize
-    ):
-        gene.maxfrag += -1
-    while True:
-        if count > len(gene.breaksites):
-            # Randomly shift a fragment
-            count = 0
-            count2 += 1
-            detectedsite = gene.rng.integers(
-                1, len(gene.breaksites) - 1, dtype=int
-            )  # dont change beginning or end
-            if gene.fragsize[detectedsite - 1] == gene.maxfrag:
-                shift = -3
-            else:
-                if gene.rng.integers(0, 2, dtype=int):
-                    shift = 3
-                else:
-                    shift = -3
-            gene.breaksites[detectedsite] = gene.breaksites[detectedsite] + shift
-            gene.fragsize = [
-                j - i for i, j in zip(gene.breaksites[:-1], gene.breaksites[1:])
-            ]
-            # if DIMPLE.dms:
-            #     tmpbreaklist = []
-            #     for idx, x in enumerate(gene.breaksites[:-1]):
-            #         if idx:
-            #             tmpbreaklist.append([x, x + gene.fragsize[idx]])
-            #         else:
-            #             tmpbreaklist.append([x + 3, x + gene.fragsize[idx] + 3])
-            #     gene.breaklist = tmpbreaklist
-            # else:
-            gene.breaklist = [
-                [x, x + gene.fragsize[idx]]
-                for idx, x in enumerate(gene.breaksites[:-1])
-            ]
-            if count2 > len(gene.breaklist) * 3:
-                gene.maxfrag += -1  # try to change for only this gene...
-                if len(gene.fragsize) * gene.maxfrag < len(gene.seq):
-                    gene = recalculate_num_fragments(gene)
-                    count = 0
-                    count2 = 0
-        count += 1
-        # Find connecting Fragments
-        if detectedsite == 0 or detectedsite == len(gene.fragsize):
-            print("Issue with primer on end of gene")
-            skip = True
-            break
-        if (
-            gene.fragsize[detectedsite] == gene.fragsize[detectedsite - 1]
-            and gene.fragsize[detectedsite] >= gene.maxfrag
-        ):
-            if all(
-                item >= gene.maxfrag for item in gene.fragsize[detectedsite + 1 :]
-            ) and not all(
-                item >= gene.maxfrag for item in gene.fragsize[: detectedsite - 1]
-            ):
-                shift = 3
-                while gene.breaksites[detectedsite] + shift in gene.problemsites:
-                    shift += 3
-            if all(
-                item >= gene.maxfrag for item in gene.fragsize[: detectedsite - 1]
-            ) and not all(
-                item >= gene.maxfrag for item in gene.fragsize[detectedsite + 1 :]
-            ):
-                shift = -3
-                while gene.breaksites[detectedsite] + shift in gene.problemsites:
-                    shift += -3
-            else:
-                if (
-                    detectedsite < len(gene.fragsize) / 2
-                ):  # should be based on problemsites not where it is located in the gene
-                    shift = 3
-                    while gene.breaksites[detectedsite] + shift in gene.problemsites:
-                        shift += 3
-                else:
-                    shift = -3
-                    while gene.breaksites[detectedsite] + shift in gene.problemsites:
-                        shift += -3
-        elif gene.fragsize[detectedsite] > gene.fragsize[detectedsite - 1]:
-            shift = 3
-            while gene.breaksites[detectedsite] + shift in gene.problemsites:
-                shift += 3
-        elif gene.fragsize[detectedsite] < gene.fragsize[detectedsite - 1]:
-            shift = -3
-            while gene.breaksites[detectedsite] + shift in gene.problemsites:
-                shift += -3
-        elif (
-            gene.fragsize[detectedsite] == gene.fragsize[detectedsite - 1]
-            and gene.fragsize[detectedsite] < gene.maxfrag
-        ):
-            shift = -3
-            while gene.breaksites[detectedsite] + shift in gene.problemsites:
-                shift = -shift
-                if shift < 0:
-                    shift += -3
-        # Process shift and reprocess fragments
-        gene.breaksites[detectedsite] = gene.breaksites[detectedsite] + shift
-        gene.fragsize = [
-            j - i for i, j in zip(gene.breaksites[:-1], gene.breaksites[1:])
-        ]
-        # if DIMPLE.dms:
-        #     tmpbreaklist = []
-        #     for idx, x in enumerate(gene.breaksites[:-1]):
-        #         if idx:
-        #             tmpbreaklist.append([x, x + gene.fragsize[idx] + 3])
-        #         else:
-        #             tmpbreaklist.append([x + 3, x + gene.fragsize[idx] + 3])
-        #     gene.breaklist = tmpbreaklist
-        # else:
-        gene.breaklist = [
-            [x, x + gene.fragsize[idx]]
-            for idx, x in enumerate(gene.breaksites[:-1])
-        ]
-        # recheck for size limit issues
-        tmpsite = [
-            topidx for topidx, item in enumerate(gene.fragsize) if item > gene.maxfrag
-        ]
-        if tmpsite:
-            # pick which side to adjust
-            if tmpsite[0] == len(gene.fragsize):
-                detectedsite = tmpsite[0]
-            elif tmpsite[0] == 0:
-                detectedsite = tmpsite[0] + 1
-            elif tmpsite[0] == detectedsite and tmpsite[0] + 1 < len(gene.fragsize):
-                detectedsite = tmpsite[0] + 1
-            else:
-                detectedsite = tmpsite[0]
-        else:
-            break
-    print(gene.fragsize)
-    # align all linked genes to the same breaksites
-    for tmp in gene.linked:
-        OLS[tmp].breaksites = gene.breaksites
-        OLS[tmp].fragsize = gene.fragsize
-        OLS[tmp].breaklist = gene.breaklist
-    return skip
-
-
-def check_overhangs(gene, OLS, overlapL, overlapR):
-    """TODO:
-    Docstring
-    """
-    # Force all overhangs to be different within a gene (no more than 2 matching in a row)
-    switched = False
-    if not isinstance(gene, DIMPLE):
-        raise TypeError("Not an instance of the DIMPLE class")
-    while True:
-        detectedsites = set()  # stores matching overhangs
-        for idx, y in enumerate(gene.breaklist):
-            overhang_F = gene.seq[
-                y[0] - DIMPLE.cutsite_overhang - overlapL: y[0] - overlapR
-            ]  # Forward overhang
-            overhang_R = gene.seq[
-                y[1] + overlapL: y[1] + DIMPLE.cutsite_overhang + overlapR
-            ]  # Reverse overhang
-            if (
-                overhang_F == overhang_R
-                or overhang_F == overhang_R.reverse_complement()
-            ):
-                detectedsites.update([idx])
-        # overhang = []
-        # for idx, y in enumerate(gene.breaklist):
-        #     overhang.append([gene.seq[y[0] - DIMPLE.cutsite_overhang - overlapL: y[0] - overlapR], idx])  # Forward overhang
-        #     overhang.append([gene.seq[y[1] + overlapL: y[1] + DIMPLE.cutsite_overhang + overlapR], idx + 1])  # Reverse overhang
-        # detectedsites = set()  # stores matching overhangs
-        # for i in range(len(overhang)):  # check each overhang for matches
-        #     for j in [x for x in range(len(overhang)) if x != i]:  # permutate over every overhang combination to find matches
-        #         #if overhang[i][0] == overhang[j][0] or overhang[i][0][:3] == overhang[j][0][:3] or overhang[i][0][1:] == overhang[j][0][1:] or overhang[i][0] == overhang[i][0].reverse_complement():  # no 3 matching bases
-        #         if overhang[i][0] == overhang[j][0] or overhang[i][0] == overhang[i][0].reverse_complement():  # no 3 matching bases
-        #             detectedsites.update([overhang[i][1]])
-        for detectedsite in detectedsites:
-            switched = True
-            if detectedsite == 0:
-                detectedsite = 1  # don't mess with the first cut site
-            print(
-                "------------------ Fragment size swapped due to matching overhangs ------------------"
-            )
-            skip = switch_fragmentsize(gene, detectedsite, OLS)
-        else:  # if no detected sites
-            break
-    return switched
 
 
 def generate_DMS_fragments(
@@ -1647,386 +1214,6 @@ def generate_DMS_fragments(
         # Record finished gene for aligned genes
         finishedGenes.extend([ii])
 
-
-def combine_fragments(tandem, num_frag_per_oligo, split):
-    """TODO:
-    Docstring
-    """
-    tandem_seq = []
-    barcodes = []
-    if split:
-        tmpF = DIMPLE.barcodeF.pop(0)
-        tmpR = DIMPLE.barcodeR.pop(0)
-    direction = -1
-    while len(tandem) > num_frag_per_oligo:
-        tmp = tandem.pop(0)
-        tmp_tandem = tmp.seq
-        tandem_id = tmp.id
-        for x in range(num_frag_per_oligo - 1):
-            if split:
-                name = tmp.id
-                tmp = tandem.pop(0)
-                if direction == 1:
-                    tmp_tandem += (
-                        "G"
-                        + DIMPLE.cutsite.reverse_complement()
-                        + tmpR
-                        + tmpF
-                        + DIMPLE.cutsite
-                        + "C"
-                        + tmp.seq
-                    )  # concatenate and add cut sites with buffer
-                    direction = -1
-                else:
-                    tmp_tandem += (
-                        "G"
-                        + DIMPLE.cutsite.reverse_complement()
-                        + tmpR
-                        + tmpF
-                        + DIMPLE.cutsite
-                        + "C"
-                        + tmp.seq.reverse_complement()
-                    )  # concatenate and add cut sites with buffer
-                    direction = 1
-                tandem_id += "+" + tmp.id
-                barcodes.append(SeqRecord(tmpR, id=name))
-                barcodes.append(SeqRecord(tmpF, id=tmp.id))
-            else:
-                tmp = tandem.pop(0)
-                if direction == 1:
-                    tmp_tandem += (
-                        "G"
-                        + DIMPLE.cutsite.reverse_complement()
-                        + "ACGT"
-                        + DIMPLE.cutsite
-                        + "C"
-                        + tmp.seq
-                    )  # concatenate and add cut sites with buffer
-                    direction = -1
-                else:
-                    tmp_tandem += (
-                        "G"
-                        + DIMPLE.cutsite.reverse_complement()
-                        + "ACGT"
-                        + DIMPLE.cutsite
-                        + "C"
-                        + tmp.seq.reverse_complement()
-                    )
-                    direction = 1
-                tandem_id += "+" + tmp.id
-        tandem_seq.append(SeqRecord(tmp_tandem, id=tandem_id, description=""))
-
-    if tandem:
-        direction = -1
-        print(len(tandem))
-        tmp = tandem.pop(0)
-        tmp_tandem = tmp.seq
-        tandem_id = tmp.id
-        while tandem:
-            if split:
-                name = tmp.id
-                tmp = tandem.pop(0)
-                if direction == 1:
-                    tmp_tandem += (
-                        "G"
-                        + DIMPLE.cutsite.reverse_complement()
-                        + "ACGT"
-                        + DIMPLE.cutsite
-                        + "C"
-                        + tmp.seq
-                    )  # concatenate and add cut sites with buffer
-                    direction = 1
-                else:
-                    tmp_tandem += (
-                        "G"
-                        + DIMPLE.cutsite.reverse_complement()
-                        + "ACGT"
-                        + DIMPLE.cutsite
-                        + "C"
-                        + tmp.seq.reverse_complement()
-                    )  # concatenate and add cut sites with buffer
-                    direction = -1
-                tandem_id += "+" + tmp.id
-                barcodes.append(SeqRecord(tmpR, id=name))
-                barcodes.append(SeqRecord(tmpF, id=tmp.id))
-            else:
-                tmp = tandem.pop(0)
-                if direction == -1:
-                    tmp_tandem += (
-                        "G"
-                        + DIMPLE.cutsite.reverse_complement()
-                        + "ACGT"
-                        + DIMPLE.cutsite
-                        + "C"
-                        + tmp.seq
-                    )  # concatenate and add cut sites with buffer
-                    direction = 1
-                else:
-                    tmp_tandem += (
-                        "G"
-                        + DIMPLE.cutsite.reverse_complement()
-                        + "ACGT"
-                        + DIMPLE.cutsite
-                        + "C"
-                        + tmp.seq
-                    )
-                    direction = -1
-                tandem_id += "+" + tmp.id
-        difference = len(tandem_seq[-1].seq) - len(tmp_tandem)
-        barF = DIMPLE.barcodeF.pop(0)
-        barR = DIMPLE.barcodeR.pop(0)
-        while difference / 2 > len(barF):
-            barF += DIMPLE.barcodeF.pop(0)
-            barR += DIMPLE.barcodeR.pop(0)
-        tmpfrag = (
-            barF.seq[0 : int(difference / 2)]
-            + tmp_tandem
-            + barR.seq.reverse_complement()[0 : difference - int(difference / 2)]
-        )
-        tandem_seq.append(SeqRecord(tmpfrag, id=tandem_id, description=""))
-        print("Partial sequence" + str(len(tmpfrag)))
-    return tandem_seq
-
-
-def print_all(OLS, folder="", config=None):
-    """Writes oligos and primers to files."""
-    if not isinstance(OLS[0], DIMPLE):
-        raise TypeError("Not an instance of the DIMPLE class")
-    alloligos = []
-    allprimers = []
-    for obj in OLS:
-        try:
-            alloligos.extend(obj.oligos)
-            allprimers.extend(obj.barPrimer)
-            allprimers.extend(obj.genePrimer)
-        except AttributeError:
-            print(obj.geneid + " has not been processed")
-    # Remove redundant sequences?
-    SeqIO.write(
-        alloligos, os.path.join(folder.replace("\\", ""), "All_Oligos.fasta"), "fasta"
-    )
-    SeqIO.write(
-        allprimers, os.path.join(folder.replace("\\", ""), "All_Primers.fasta"), "fasta"
-    )
-
-
-def post_qc(OLS, config=None):
-    if config is not None:
-        DIMPLE.enzyme = getattr(config, "enzyme", DIMPLE.enzyme)
-    logger.info("Running post QC")
-    if not isinstance(OLS[0], DIMPLE):
-        raise TypeError("Not an instance of the DIMPLE class")
-    # Post QC
-    all_oligos = []
-    all_barPrimers = []
-    for obj in OLS:
-        logger.info(f'Running QC for {obj.geneid}')
-        try:
-            all_oligos.extend(obj.oligos)
-            all_barPrimers.extend(obj.barPrimer)
-        except AttributeError:
-            print(obj.geneid + " has not been processed")
-
-        if DIMPLE.enzyme is not None:
-            logger.info(f"Checking oligo assembly for {obj.geneid}")
-            check_final_assembly(obj)
-        else:
-            logger.info(f"Skipping oligo assembly check for {obj.geneid} as no enzyme is specified.")
-
-
-
-    print("Running QC for barcode primer specificity")
-    cassetteSet = set(all_oligos[0].id[:-6])
-    uCassette = [SeqRecord(all_oligos[0].seq, id=all_oligos[0].id[:-6])]
-    for idx in range(len(all_oligos)):
-        if not all_oligos[idx].id[:-6] in cassetteSet:
-            uCassette.append(SeqRecord(all_oligos[idx].seq, id=all_oligos[idx].id[:-6]))
-        cassetteSet.add(all_oligos[idx].id[:-6])
-    grouped = iter(all_barPrimers)
-    grouped = zip(grouped, grouped)  # create the combinatorial comparisons
-    nonspecific = []
-    # iterate over every barcode primer pair and match to each oligo to check for nonspecific amplification
-    for idxPrime, primers in enumerate(
-        grouped
-    ):  # iterate over every barcode primer pair
-        print("Checking primer set:" + primers[0].id[:-2])
-        for idxCassette, fragment in enumerate(
-            uCassette
-        ):  # iterate over every OLS oligo
-            if (
-                primers[0].id.split("_")[2] != all_oligos[idxCassette].id.split("_")[2]
-            ):  # ignore designed annealing (same name)
-                fragname = fragment.id
-                fragment = fragment.seq
-                non = [[False], [False]]
-                for idxDirection, primer in enumerate(primers):
-                    primername = primer.id
-                    primer = primer.seq
-                    for i in range(
-                        len(fragment) - len(primer)
-                    ):  # iterate over the length of the oligo for a binding site
-                        match = [
-                            primer[j].lower() == fragment[i + j].lower()
-                            for j in range(len(primer))
-                        ]
-                        first = 10
-                        for k in range(len(match) - 3):
-                            if (match[k] and match[k + 1] and match[k + 3]) or (
-                                match[k] and match[k + 1] and match[k + 2]
-                            ):
-                                first = k
-                                break
-                        if (
-                            sum(match[first:]) > len(primer[first:]) * 0.8
-                            and sum(match[first:]) > 6
-                            and match[-1]
-                        ):  # string compare - sum of matched nt is greater than 80%
-                            try:
-                                melt = mt.Tm_NN(
-                                    primer[first:],
-                                    c_seq=fragment[
-                                        i + first : i + len(primer)
-                                    ].complement(),
-                                    nn_table=mt.DNA_NN2,
-                                    de_table=mt.DNA_DE1,
-                                    imm_table=mt.DNA_IMM1,
-                                )
-                                # if melt>20:
-                                #     print('Found non-specific match:'+fragment[i:i+len(primer)])
-                                #     print('                 primer:'+primer+' Tm:'+str(round(melt,1)))
-                                if melt > 35:
-                                    non[0].append(True)
-                            except ValueError as valerr:
-                                # print(str(valerr)+". Please check position manually:"+str(i+1)+" forward")
-                                # print('Primer:'+primer)
-                                # print('Match: '+fragment[i:i+len(primer)])
-                                pass
-                    fragment = fragment.reverse_complement()
-                    for i in range(len(fragment) - len(primer)):
-                        match = [
-                            primer[j].lower() == fragment[i + j].lower()
-                            for j in range(len(primer))
-                        ]
-                        first = 10
-                        for k in range(0, len(match) - 3, 1):
-                            if match[k] and match[k + 1] and match[k + 3]:
-                                first = k
-                                break
-                        if (
-                            sum(match[first:]) > len(primer[first:]) * 0.8
-                            and sum(match[first:]) > 6
-                            and match[-1]
-                        ):  # string compare - sum of matched nt is greater than 80%
-                            try:
-                                melt = mt.Tm_NN(
-                                    primer[first:],
-                                    c_seq=fragment[
-                                        i + first : i + len(primer)
-                                    ].complement(),
-                                    nn_table=mt.DNA_NN2,
-                                    de_table=mt.DNA_DE1,
-                                    imm_table=mt.DNA_IMM1,
-                                )
-                                # if melt > 20:
-                                #     print('Found non-specific match:'+fragment[i:i+len(primer)])
-                                #     print('                 primer:'+primer+' Tm:'+str(melt))
-                                if melt > 35:
-                                    non[1].append(True)
-                            except ValueError as valerr:
-                                # print(str(valerr)+". Please check position manually:"+str(i+1)+" reverse")
-                                # print('Primer:'+primer)
-                                # print('Match: '+fragment[i:i+len(primer)])
-                                pass
-                    if sum(non[0]) == 0 and sum(non[1]) == 0:
-                        break
-                    if sum(non[0]) > 0 and sum(non[1]) > 0:
-                        nonspecific.append([primername, fragname])
-                        print("Found Non-specific Amplification")
-    if nonspecific:
-        print("Nonspecific Primers: (Manually changing primer sequence recommended)")
-        print(nonspecific)
-    else:
-        print("No non-specific primers detected")
-
-def check_final_assembly(gene, config=None):
-    """Test that each oligo assembles properly and contains the designed mutation."""
-
-    if config is not None:
-        DIMPLE.enzyme = getattr(config, "enzyme", DIMPLE.enzyme)
-
-    # Check whether the enzyme is set.
-    if DIMPLE.enzyme:
-        if DIMPLE.enzyme == "BsmBI":
-            enzyme = BsmBI
-        elif DIMPLE.enzyme == "BsaI":
-            enzyme = BsaI
-        else:
-            logger.warning("Enzyme not recognized. Not performing assembly check.")
-            return None
-    else:
-        logger.warning("No enzyme set. Not performing assembly check.")
-        return None
-
-    n_fragments = len(gene.genePrimer) // 2
-    full_template = Dseqrecord(gene.seq, circular=True)
-
-    backbones = []
-    oligo_primer_dseqs = []
-    logger.info(f"Testing assembly for {gene.geneid}")
-    logger.info(f"Using enzyme: {DIMPLE.enzyme}")
-    logger.info(f"Number of fragments: {n_fragments}")
-
-    for frag in range(0, n_fragments):
-        fwd_primer = Dseqrecord(gene.genePrimer[frag * 2])
-        rev_primer = Dseqrecord(gene.genePrimer[frag * 2+1])
-        template_pcr_product = Dseqrecord(pcr(fwd_primer, rev_primer, full_template))
-        cut_template_product = max(template_pcr_product.cut(enzyme), key = len)
-        backbones.append(cut_template_product)
-
-        fwd_oligo_primer = Dseqrecord(gene.barPrimer[frag * 2])
-        rev_oligo_primer = Dseqrecord(gene.barPrimer[frag * 2+1])
-        oligo_primer_dseqs.append((fwd_oligo_primer, rev_oligo_primer))
-
-    for variant in gene.designed_variants:
-        variant_dict = gene.designed_variants[variant]
-        # Get the fragment that the variant is in.
-        fragment = variant_dict['fragment']
-        sequence = variant_dict['xfrag']
-        oligo_sequence = variant_dict['oligo_sequence']
-        # Simulate PCR of oligo with oligo primers.
-        fwd_oligo_primer, rev_oligo_primer = oligo_primer_dseqs[fragment-1]
-        oligo_pcr_product = Dseqrecord(pcr(fwd_oligo_primer, rev_oligo_primer, oligo_sequence))
-
-        try:
-            cut_oligo_product = max(oligo_pcr_product.cut(enzyme), key = len)
-        except ValueError as error:
-            logger.error(f"Oligo cut site issue: {variant}")
-            logger.error(str(error))
-
-        try:
-            assembled = (cut_oligo_product + backbones[fragment-1]).looped()
-            if str(sequence[4:-4]) not in str(assembled.seq):
-                logger.error(f"Assembly product incorrect: {variant}.")
-                logger.error(f"Expected variant to contain: {str(sequence[4:-4])}")
-                logger.error(f"Predicted assembly product: {str(assembled.seq)}")
-
-        except TypeError as error:
-            logger.error(f"Oligo does not assemble with template: {variant}")
-            logger.error(str(error))
-
-
-# Incremental module decomposition: keep legacy implementations available while
-# routing public symbols through focused modules.
-_legacy_find_geneprimer = find_geneprimer
-_legacy_find_fragment_primer = find_fragment_primer
-_legacy_check_nonspecific = check_nonspecific
-_legacy_recalculate_num_fragments = recalculate_num_fragments
-_legacy_switch_fragmentsize = switch_fragmentsize
-_legacy_check_overhangs = check_overhangs
-_legacy_combine_fragments = combine_fragments
-_legacy_print_all = print_all
-_legacy_post_qc = post_qc
-_legacy_test_final_assembly = check_final_assembly
 
 from DIMPLE.fragment_layout import (  # noqa: E402
     check_overhangs,
